#802 Number generator button next to numeric control
