+++ canvas_modules/common-canvas/src/common-properties/controls/numberfield/numberfield.scss
@@ -20,6 +20,23 @@
 		margin-right: 0;
 		margin-bottom: 0;
 	}
+	.properties-number-generator {
+		margin-left: $spacing-05;
+		right: $spacing-01;
+		// Button dimensions 40x40 px
+		width: $spacing-08;
+		min-height: $spacing-08;
+		padding-left: $spacing-04;
+		padding-right: $spacing-04;
+	}
+	&.numberfield-with-number-generator {
+		width: 100%;
+		display: inline-flex;
+		align-items: flex-end;
+		&.error, &.warning {
+			align-items: center;
+		}
+	}
 }
 
 .properties-table-cell-control { // overrides when in a table cell
